feat: send mail with nodemailer and gmail - wip
