+++ .env.example
@@ -4,9 +4,6 @@ NODE_ENV=development
 
 # SMTP SERVICE CONFIG FOR NODEMAILER
 SMTP_SERVER_SERVICE='service'
-SMTP_SERVER_HOST='host'
-SMTP_SERVER_PASS='pass'
-SITE_MAIL_RECIEVER='email'
 
 # OAUTH2 GMAIL CREDENTIALS
 GMAIL_CLIENT_ID="id"
