+++ app/components/ui/Carousels/Buttons/SelectedSnapDisplay.tsx
