+++ app/services/sendEmail.tsx
@@ -1,55 +1,29 @@
-'use server';
-import nodemailer from 'nodemailer';
-import { getOAuthToken } from './generateToken';
+import type { EmailData } from '@/types/data/EmailDataProps';
 
-const transporter = nodemailer.createTransport({
-  service: process.env.SMTP_SERVER_SERVICE,
-  host: process.env.SMTP_SERVER_HOST,
-  port: 465,
-  secure: true,
-  auth: {
-    type: 'OAuth2',
-    user: process.env.SMTP_SERVER_USERNAME,
-    clientId: process.env.GMAIL_CLIENT_ID,
-    clientSecret: process.env.GMAIL_CLIENT_SECRET,
-    refreshToken: process.env.GMAIL_REFRESH_TOKEN,
-    accessToken: await getOAuthToken(),
-  },
-});
-
-export const sendEmail = async ({
-  email,
-  sendTo,
-  subject,
-  text,
-  html,
-}: {
-  email: string;
-  sendTo: string;
-  subject: string;
-  text: string;
-  html?: string;
-}) => {
+export const sendEmail = async (
+  data: EmailData
+): Promise<{ success: boolean; messageId?: string }> => {
   try {
-    await transporter.verify();
-  } catch (error) {
-    console.error(
-      'Something went wrong',
-      process.env.SMTP_SERVER_USERNAME,
-      process.env.SMTP_SERVER_SERVICE,
-      error
-    );
-
-    const info = await transporter.sendMail({
-      from: email,
-      to: process.env.SITE_MAIL_RECIEVER,
-      subject,
-      text,
-      html: html ? html : '',
+    const response = await fetch('/api/contact', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        email: data.email,
+        subject: data.subject,
+        message: data.text,
+      }),
     });
 
-    console.log('Message sent: ', info.messageId);
-    console.log('Mail sent to: ', sendTo);
-    return info;
+    if (!response.ok) {
+      const errorData = await response.json();
+      throw new Error(errorData.error || "Erreur lors de l'envoi");
+    }
+
+    return await response.json();
+  } catch (error) {
+    console.error("Erreur lors de l'envoi:", error);
+    throw error;
   }
 };
