+++ app/types/data/EmailDataProps.tsx
@@ -1,0 +1,6 @@
+export interface EmailData {
+  email: string;
+  sendTo: string;
+  subject: string;
+  text: string;
+}
