+++ components.json
@@ -11,11 +11,11 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "lib/components",
-    "utils": "lib/utils",
-    "ui": "lib/components/ui",
-    "lib": "lib",
-    "hooks": "lib/hooks"
+    "components": "@/lib/components",
+    "utils": "@/lib/utils",
+    "ui": "@/lib/components/ui",
+    "lib": "@/lib",
+    "hooks": "@/lib/hooks"
   },
   "iconLibrary": "lucide"
 }
