--- docs/todo.md
@@ -1,70 +1,0 @@
-# TODO
-
-## MVP
-
-**Carte présentation**
-
-- [x] Photo
-- [x] Nom et prénom
-- [x] Présentation (1 phrase)
-
-**Carte localisation**
-
-- [x] Localisation sur une carte (api externe)
-- [x] Mobilité dans la carte
-- [x] Animation de la France à ma localisation
-
-**Carte CV**
-
-- [x] Titre
-- [x] Année d'actualisation
-- [x] Expériences développeur
-- [x] Autres expériences pertinentes
-- [x] Lecture du CV
-- [x] Téléchargement du CV
-
-**Carte projets**
-
-- [x] Titre "Projets"
-- [x] Liste de 3 projets
-- [x] Description des projets
-- [x] Liens vers les projets
-- [x] Liens vers les dépôts
-- [x] Liens vers les articles
-- [x] Technologies utilisées
-- [x] Images des projets
-
-**Carte compétences**
-
-- [x] Titre "Compétences"
-- [x] Liste de compétences
-- [x] Carousel de compétences
-
-**Carte recommandations**
-
-- [x] Titre
-- [x] Liste de recommandations
-- [x] Auteur de la recommandation
-- [x] Contenu de la recommandation
-- [x] Slider de recommandations
-- [x] Avatar de l'auteur
-- [x] LinkedIn de l'auteur
-
-**Carte contact**
-
-- [x] Titre
-- [x] Formulaire
-- [x] Identité de l'expéditeur (nom, prénom, email)
-- [x] Message
-- [x] Envoi du message
-- [x] Message d'erreur
-  - [x] Nom
-  - [x] Mail
-- [x] Message de succès
-- [x] Envoi du message par email
-- [x] Proposition de contact par téléphone
-
-**Liens réseaux sociaux**
-
-- [x] Github
-- [x] Linkedin
