+++ tsconfig.json
@@ -19,10 +19,11 @@
     ],
     "baseUrl": "app/",
     "paths": {
-      "@/components/*": ["_components/*"],
-      "@/ui/*": ["_components/ui/*"],
-      "@/layout/*": ["_components/layout/*"],
-      "@api/*": ["pages/api/*"],
+      "@/components/*": ["components/*"],
+      "@/forms/*": ["components/forms/*"],
+      "@/ui/*": ["components/ui/*"],
+      "@/layout/*": ["components/layout/*"],
+      "@/features/*": ["features/*"],
       "@/config/*": ["config/*"],
       "@/fonts/*": ["fonts/*"],
       "@/lib/*": ["lib/*"],
@@ -31,7 +32,8 @@
       "@/utils/*": ["utils/*"],
       "@/hooks/*": ["hooks/*"],
       "@/schemas/*": ["schemas/*"],
-      "@/services/*": ["services/*"]
+      "@/services/*": ["services/*"],
+      "@/api/*": ["api/*"]
     },
     "target": "ES2017"
   },
